==> mtd/code/workstation/obtaining-experimental-info/1_Get_Series_Info.ipynb <==
from Bio import Entrez
import pandas as pd
import re

# 指定Entrez使用的邮件地址，这是必需的
Entrez.email = "3575318018@qq.com"

# 保存表格的名称开头，便于分类
# key_name = "An"
# name = "Aspergillus niger"
# key_name = "Tr"
# name = "Trichoderma reesei"
key_name = "Nc"
name = "Neurospora crassa"

# 指定搜索的关键词和数据库
search_term = name + " AND GSE[Entry Type]"
database = "gds"

# 使用Entrez进行搜索并获取ID列表
handle1 = Entrez.esearch(db=database, term=search_term, retstart=0, retmax=200)
record = Entrez.read(handle1)
id_list = record["IdList"]

# 获取所有ID的详细信息
handle2 = Entrez.esummary(db=database, id=",".join(id_list))
records = Entrez.read(handle2)
records = list(records)

def get_gse_info(records):
    """根据records,获取相关菌种的gse数据集关键信息.
    """
    gse_list = []
    for record in records:
        title = record["title"]
        summary = record["summary"]
        Gse_id = record['Accession']
        samples_num = record['n_samples']
        
        if "PubMedIds" in record and len(record["PubMedIds"]) > 0:
            pmid = str(record["PubMedIds"][0])
            match = re.search(r'\d+', pmid)
            if match:
                pmid = match.group(0)
            else:
                pmid = "N/A"
        else:
            pmid = "N/A"

        gse_list.append({'GSE Series':Gse_id,'Title':title,'Summary':summary,'Samples':samples_num,'PMID':pmid})

    df_gse = pd.DataFrame(gse_list)
    df_gse.to_csv(key_name + '_' + 'GSE_Info_All.csv', index=False, columns=['GSE Series','Title','Summary','Samples','PMID'])  # All代表是NCBI上记录的此菌种的所有GSE数据集
    
    # 将df_gse中GSE Series的GSE号保存在一个txt文件中，行尾序列为LF格式
    with open(key_name + '_' + 'All_GSE.txt', 'w') as f:
        for i in df_gse['GSE Series']:
            f.write(i + '\n')

    return df_gse


def get_gsm_and_sample_name(records):
    """根据records,获取 GSM Sample ID 与 name 及样本对应的GSE Series ID.
    """
    sample_list = []
    for record in records:
        Gse_id = record['Accession']
        samples = record['Samples']
        for key in samples:
            accession = key['Accession']
            title = key['Title']
            sample_list.append({'GEO Series': Gse_id,'Accession': accession,'Title': title})
            
    df = pd.DataFrame(sample_list)
    df.to_csv(key_name + '_' + 'GSM_ID.csv', index=False, columns=['GEO Series','Accession','Title'])
    
    return df

get_gse_info(records)

get_gsm_and_sample_name(records)

# for record in records:
#     Samples_name = record['Samples']
#    #  for i in Samples_name:
#       #  Samples_name_singel = Samples_name['Accession']
#       #  Samples_name_title = Samples_name['Title']

#     print("Samples_name: {}".format(Samples_name))

# print("Title: {}".format(title))
# print("Summary: {}".format(summary))
# print("Pmid: {}".format(pmid))
# print("GSE: {}".format(Gse_id))
# print("Samples: {}".format(samples_num))

# for record in records:
#     samples = record['Samples']
#     for sample in samples:
#         accession = sample['Accession']
#         title = sample['Title']
#         print("Accession: {}, Title: {}".format(accession, title))
