==> mtd/code/workstation/obtaining-experimental-info/4_Parsing_GSE_Soft_File.ipynb <==
### 解析GSE.soft文件

import re
import pandas as pd
import os

def extract_information(file_path):
    """从GSE的soft文件中提取提交单位与国家
    """

    information = []

    with open(file_path) as file:
        content = file.read()

    # 使用正则表达式进行匹配提取信息
    accession_pattern = r'!Series_geo_accession = (\w+)'
    accessions = re.findall(accession_pattern, content)
    information.append(accessions[0])

    institute_pattern = r'!Series_contact_institute = (.*)'
    institutes = re.findall(institute_pattern, content)
    information.append(institutes[0])

    country_pattern = r'!Series_contact_country = (.*)'
    countries = re.findall(country_pattern, content)
    information.append(countries[0])

    organism_pattern = r'!Series_platform_organism = (.*)'
    organism = re.findall(organism_pattern, content)
    information.append(organism[0])

    return information

单个测试

# location = './Mt/GSE_soft/GSE110062_GSE.soft'
# extract_information(location)

### 解析下载的GSE，只保留研究核心菌种是目标菌种的GSE


path = './Nc/GSE_soft/'
key_name = 'Nc'
results = []

for filename in os.listdir(path):
    result = extract_information(os.path.join(path, filename))
    results.append(result)

df = pd.DataFrame(results, columns=['GEO Series ID', 'Institute', 'Country', 'Organism'])
display(df)

df.to_csv(key_name + '_soft.csv', index=False )

### 读取菌种在NCBI上有相应文献的所有的GSE

df1 = pd.read_csv(key_name + '_GSE_Info_All.csv').copy()

df1.rename(columns={'Title': 'Study Title'}, inplace=True)

# df1取GEO Series ID列，Samples列，PMID列
df1 = df1[['GSE Series', 'Samples', 'PMID', 'Study Title']]

# PMID列处理为整数
df1['PMID'] = df1['PMID'].fillna(0)
df1['PMID'] = df1['PMID'].astype(int)

df1

df2 = pd.read_csv(key_name + '_GSE_Info_Pub.csv').copy()

# PMID列处理为整数
df2['PMID'] = df2['PMID'].astype(int)

df2

df3 = pd.read_csv(key_name + '_soft.csv').copy()

# GEO Series ID列更名为GSE Series
df3 = df3.rename(columns={'GEO Series ID': 'GSE Series'})

df3

# df1与df2合并，依据PMID列, 以df1为主
df4 = pd.merge(df1, df2, on='PMID', how='left').copy()

# df去掉Country列
df4 = df4.drop(columns=['Country'])
df4 = df4.drop(columns=['Keywords'])

# df4中每一列的空值替换为空
df4 = df4.fillna('')

# df4中的Abstract列删除BACKGROUND:
df4['Abstract'] = df4['Abstract'].str.replace('BACKGROUND:', '')

# df4中的Yead列显示的是年份+月份+日期，只保留年份
df4['Year'] = df4['Year'].str[:4]

df4

# df4与df3合并，依据GEO Series列, 以df4为主
df5 = pd.merge(df4, df3, on='GSE Series', how='left')
df5

# df5列顺序调整
df5 = df5[['GSE Series', 
           'Samples', 
           'Study Title', 
           'Title',
           'Abstract',
           'Organism',
           'Country',
           'Institute', 
           'Journal',
           'Year',
           'PMID' 
           ]]

# GSE Series列更名为Series Accession
df5 = df5.rename(columns={'GSE Series': 'Series Accession'})

df5

只保留Soft文件对应的数据集（具体那些GSE需要先去NCBI上看看）

# 删除df中Organism为NaN的行
df6 = df5.dropna(subset=['Organism']) # 代表不是我们挑的数据集
df6

# 如果df6中PMID列的值为0，则删除该行
df7 = df6[df6['PMID'] != 0] # 代表没对应文献
df7

# 重置索引
df6.reset_index(drop=True, inplace=True)

# df5保存为csv
df6.to_csv(key_name + '_dataset.csv', index=False)
