==> mtd/code/workstation/obtaining-experimental-info/6_Add_Run.ipynb <==
import pandas as pd

key_name = 'Tr'

# 读取mtd/code/Obtaining_experimental_information/SraRunInfo.csv
df_ncbi = pd.read_csv(key_name + '_SraRunInfo.csv').copy()
df_ncbi

df1 = df_ncbi[['Run', 'Experiment']].copy()

# 如果Experiment列存在重复，则将重复的行合并
df_ncbi = df_ncbi.groupby('Experiment').agg({
    'Run':lambda x: ','.join(x),
    }).reset_index().copy( )

df_ncbi

# 保存为tsv
df_ncbi.to_csv (key_name + '_SraRunInfo_Run.csv', sep=',', index=False)

df_nc_sample = pd.read_csv(key_name + '_sample.csv').copy()

# 将空值填充为空
df_nc_sample = df_nc_sample.fillna('')

# 去掉Experiment列中的值的第一个空格
df_nc_sample['Experiment'] = df_nc_sample['Experiment'].apply(lambda x: x.strip())

df_nc_sample = df_nc_sample[df_nc_sample['Library Strategy'] == 'RNA-Seq']

df_nc_sample

# 如果df_ncbi的Experiment列中的值与df_nc_sample的Experiment列中的值相同，则将df_ncbi的Run列中的值赋给df_nc_sample的Run
df_nc_sample['Run'] = df_nc_sample['Experiment'].map(df_ncbi.set_index('Experiment')['Run'])
df_nc_sample

#### 有些数据没Run，即未公开，删除掉

df = df_nc_sample.dropna(subset=['Run'])
df

### 为什么样本数减少了？
- 有的样本没公开
- 有的样本不是该菌种

# 保存为csv
df.to_csv (key_name + '_sample.csv', sep=',', index=False)

common_experiments = set(df_nc_sample['Experiment']).intersection(set(df_ncbi['Experiment']))
missing_experiments = set(df_nc_sample['Experiment']) - common_experiments
print("Number of common experiments:", len(common_experiments))
print("Number of missing experiments:", len(missing_experiments))
