==> mtd/code/workstation/obtaining-experimental-info/7_Check.ipynb <==
import pandas as pd

key_name = 'Tr'

df1 = pd.read_csv(key_name + '_sample.csv', sep=',')

# df1只保留Experiment与Run两列
df1 = df1[['Experiment', 'Run']]

df1

df1['Run'] = df1['Run'].str.split(',')
df1 = df1.explode('Run')

df1

# df1保存为tsv
df1.to_csv(key_name + '_核对.csv', sep=',', index=False)

df1
