==> mtd/code/workstation/obtaining-experimental-info/8_Sample_id.ipynb <==
import pandas as pd 

key_name = 'Tr'
df = pd.read_csv(key_name + '_sample.csv').copy()
df

# df创建新列
df["Carbon Sources"] = ""
df["Temperature"] = ""
df["Time"] = ""
df["id_1"] = ""
df["id_2"] = ""
df["id_3"] = ""

# df重新排序
df = df[
    [
        "id_1",
        "id_2",
        "id_3",
        "Sample ID",
        "Sample Name",
        "Experiment",
        "Run",
        "Series Accession",
        "Study Title",
        "Species",
        "Model",
        "Library Seletion",
        "Library Source",
        "Library Strategy",
        "Samples",
        "Strain",
        "Carbon Sources",
        "Temperature",
        "Time",
        "Treatment",
        "Condition",
        "Condition_2",
        "Condition_3",
        "Year",
    ]
]

df

# df保存为excel，去掉索引
df.to_excel(key_name + '_sample_raw.xlsx', index=False)
