==> mtd/code/workstation/obtaining-experimental-info/9_Fname命名.ipynb <==
import pandas as pd
import numpy as np

# 读取当前目录下的，Nc_sample.xlsx文件，第一个表
df = pd.read_excel('Nc_sample.xlsx', sheet_name=0).copy()
df.head()

# 将id_2列，设置为空
df['id_2'] = np.nan
df

current_id = 1
last_series = df['Series Accession'].iloc[0]

# 初始化空的ID列表
id_list = []

# 遍历每一行
for series in df['Series Accession']:
    # 如果当前序列和上一个序列相同，id不变
    if series == last_series:
        id_list.append(current_id)
    # 如果不同，id加一
    else:
        current_id += 1
        id_list.append(current_id)
    # 更新上一个序列
    last_series = series

# 将ID列表添加到df中
df['id_2'] = id_list

df

# df保存为excel
df.to_excel('name.xlsx', sheet_name='Sheet1')
