==> mtd/code/workstation/obtaining-experimental-info/10_Fname.ipynb <==
import pandas as pd

path = './'
key_name = 'An'

# 读取An_sample.xlsx中的数据
df = pd.read_excel(path + key_name + '_sample_raw.xlsx', sheet_name='Sheet1').copy()
df

# 将'id_2'和'id_3'列的值填充成3位数字形式
df['id_2'] = df['id_2'].apply(lambda x: f"{x:03d}")
df['id_3'] = df['id_3'].apply(lambda x: f"{x:03d}")

df

# df中Sample ID列的值为id_1 + id_2 + id_3这三列的值加起来
df['Sample ID'] = df['id_1'] + df['id_2'] + df['id_3']
df

# 保存df，为excel
df.to_excel('Nc_sample_raw.xlsx', index=False)
